+++ ui/packages/shared/components/src/ParcaContext/index.tsx
@@ -21,6 +21,15 @@ import Spinner from '../Spinner';
 interface Props {
   loader: ReactNode;
   noDataPrompt: ReactNode;
+  profileExplorer: {
+    PaddingX: number;
+    metricsGraph: {
+      maxHeightStyle: {
+        default: string;
+        compareMode: string;
+      };
+    };
+  };
   perf?: {
     onRender?: ProfilerOnRenderCallback;
     markInteraction: (interactionName: string, sampleCount: number | string | bigint) => void;
@@ -31,6 +40,15 @@ interface Props {
 export const defaultValue: Props = {
   loader: <Spinner />,
   noDataPrompt: <NoDataPrompt />,
+  profileExplorer: {
+    PaddingX: 58,
+    metricsGraph: {
+      maxHeightStyle: {
+        default: 'calc(47vw - 24px)',
+        compareMode: 'calc(23.5vw - 24px)',
+      },
+    },
+  },
   perf: {
     onRender: () => {},
     markInteraction: () => {},