+++ ui/packages/shared/store/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.16.91 (2023-07-14)
+
+**Note:** Version bump only for package @parca/store
+
 ## [0.16.90](https://github.com/parca-dev/parca/compare/@parca/store@0.16.89...@parca/store@0.16.90) (2023-07-08)
 
 **Note:** Version bump only for package @parca/store